documentation and namespace changes
