+++ Packages/io.chainsafe.web3-unity.ramp/Runtime/Scripts/RampServiceAdapter.cs
@@ -1,4 +1,4 @@
-using ChainSafe.Gaming.UnityPackage.Common;
+using ChainSafe.Gaming.UnityPackage.Connection;
 using ChainSafe.Gaming.Web3.Build;
 using UnityEngine;
