--- Packages/io.chainsafe.web3-unity/Runtime/Common/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b3347b82b502e4f4fa3d94b8ae6c144a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
