+++ Packages/io.chainsafe.web3-unity/Runtime/Scripts/Connection/ConnectionProviderConfig.cs.meta
