+++ Packages/io.chainsafe.web3-unity/Runtime/Scripts/Connection/WalletConnectConnectionProvider.cs
@@ -4,8 +4,11 @@ using ChainSafe.Gaming.Web3.Build;
 using ChainSafe.Gaming.Web3.Evm.Wallet;
 using UnityEngine;
 
-namespace ChainSafe.Gaming
+namespace ChainSafe.Gaming.UnityPackage.Connection
 {
+    /// <summary>
+    /// WalletConnect connection provider used for connecting to a wallet using WalletConnect.
+    /// </summary>
     public class WalletConnectConnectionProvider : ConnectionProvider
     {
         [SerializeField] private WalletConnectConfigSO walletConnectConfig;
