+++ Packages/io.chainsafe.web3-unity/Runtime/Scripts/UI/ErrorOverlay.cs
@@ -5,8 +5,11 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
-namespace ChainSafe.Gaming
+namespace ChainSafe.Gaming.UnityPackage.UI
 {
+    /// <summary>
+    /// Error overlay used for displaying error messages.
+    /// </summary>
     public class ErrorOverlay : MonoBehaviour
     {
         [SerializeField] private TextMeshProUGUI errorText;
@@ -17,6 +20,10 @@ namespace ChainSafe.Gaming
             closeButton.onClick.AddListener(Close);
         }
 
+        /// <summary>
+        /// Display error messages.
+        /// </summary>
+        /// <param name="message">Error message to display.</param>
         public void DisplayError(string message)
         {
             gameObject.SetActive(true);
@@ -24,7 +31,10 @@ namespace ChainSafe.Gaming
             errorText.SetText(message);
         }
         
-        public void Close()
+        /// <summary>
+        /// Close error overlay.
+        /// </summary>
+        private void Close()
         {
             gameObject.SetActive(false);
         }
