+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK HyperPlay/1.0.0/HyperPlay Samples/Scripts/HyperPlayLoginProvider.cs
@@ -1,9 +1,8 @@
 using ChainSafe.Gaming.HyperPlay;
-using ChainSafe.Gaming.UnityPackage.Common;
+using ChainSafe.Gaming.UnityPackage.Connection;
 using ChainSafe.Gaming.Web3.Build;
 using ChainSafe.Gaming.Web3.Evm.Wallet;
 using Microsoft.Extensions.DependencyInjection;
-using Scenes;
 using UnityEngine;
 using UnityEngine.UI;
