+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK Lootboxes/2.6.0/Lootboxes Samples/Scripts/LootboxesServicesAdapter.cs
@@ -1,5 +1,5 @@
 using ChainSafe.Gaming.Lootboxes.Chainlink;
-using ChainSafe.Gaming.UnityPackage.Common;
+using ChainSafe.Gaming.UnityPackage.Connection;
 using ChainSafe.Gaming.Web3.Build;
 using Microsoft.Extensions.DependencyInjection;
 using UnityEngine;
