+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK Web3Auth/2.6.0/Web3Auth Samples/Scripts/Web3AuthLoginProvider.cs
@@ -5,12 +5,11 @@ using System.Runtime.InteropServices;
 using AOT;
 using System.Threading.Tasks;
 using ChainSafe.Gaming.UnityPackage;
-using ChainSafe.Gaming.UnityPackage.Common;
+using ChainSafe.Gaming.UnityPackage.Connection;
 using ChainSafe.Gaming.Web3.Analytics;
 using ChainSafe.Gaming.Web3.Build;
 using ChainSafe.GamingSdk.Web3Auth;
 using Nethereum.Hex.HexTypes;
-using Scenes;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
