+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.6.0/Web3.Unity Samples/Scripts/Prefabs/Logout.cs
@@ -1,8 +1,5 @@
-using System.Threading.Tasks;
 using ChainSafe.Gaming.UnityPackage;
-using ChainSafe.Gaming.Wallets;
-using ChainSafe.Gaming.UnityPackage.Common;
-using Scenes;
+using ChainSafe.Gaming.UnityPackage.Connection;
 using UnityEngine;
 using UnityEngine.SceneManagement;
