+++ src/UnitySampleProject/Assets/Samples/web3.unity SDK/2.6.0/Web3.Unity Samples/Scripts/Scenes/ConnectToWallet.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using ChainSafe.Gaming;
+using ChainSafe.Gaming.UnityPackage.UI;
 using UnityEngine;
 using UnityEngine.UI;
